+++ src/tx/memory.rs
@@ -21,7 +21,7 @@ impl TransactionStore for MemoryTransactionCache {
     async fn add_transaction(
         &mut self,
         transaction: &Transaction,
-        height: Option<usize>,
+        height: Option<u32>,
         hash: &BlockHash,
     ) -> Result<(), TransactionStoreError> {
         self.transactions.push(IndexedTransaction {
